move test files in test directory
