+++ packages/puppeteer-extra-plugin-adblocker/package.json
@@ -78,7 +78,7 @@
     "environmentVariables": {
       "TS_NODE_COMPILER_OPTIONS": "{\"module\":\"commonjs\"}"
     },
-    "files": ["src/*.test.ts"],
+    "files": ["test/*.test.ts"],
     "extensions": ["ts"],
     "require": ["ts-node/register"]
   },
