+++ packages/puppeteer-extra-plugin-adblocker/test/index.test.ts
@@ -1,6 +1,6 @@
 import test from 'ava'
 
-import AdblockerPlugin from './index'
+import AdblockerPlugin from '../src'
 
 const PUPPETEER_ARGS = ['--no-sandbox', '--disable-setuid-sandbox']
