+++ packages/puppeteer-extra-plugin-recaptcha/package.json
@@ -82,7 +82,7 @@
     "environmentVariables": {
       "TS_NODE_COMPILER_OPTIONS": "{\"module\":\"commonjs\"}"
     },
-    "files": ["src/*.test.ts"],
+    "files": ["test/*.test.ts"],
     "extensions": ["ts"],
     "require": ["ts-node/register"]
   },
