+++ packages/puppeteer-extra-plugin-recaptcha/test/index.test.ts
