+++ packages/puppeteer-extra-plugin-recaptcha/tsconfig.json
@@ -22,6 +22,6 @@
     "types": ["node"]
   },
   "include": [
-    "./src/**/*.ts",
+    "./src/**/*.ts"
   ]
 }
