+++ packages/puppeteer-extra-plugin/package.json
@@ -77,7 +77,7 @@
     "environmentVariables": {
       "TS_NODE_COMPILER_OPTIONS": "{\"module\":\"commonjs\"}"
     },
-    "files": ["src/**/*.test.ts"],
+    "files": ["test/*.test.ts"],
     "extensions": ["ts"],
     "require": ["ts-node/register"]
   },
